safari-600.7-branch/LayoutTests/plugins/npruntime/valueof-expected.txt

This tests that valueOf() works correctly on NPObject.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS plugin.valueOf() is 123456789
PASS plugin == 123456789 is true
PASS successfullyParsed is true

TEST COMPLETE
